⚡️ App render
