+++ src/main.tsx
@@ -1,40 +1,21 @@
-import { RouterProvider, createRouter } from "@tanstack/react-router";
 import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
+import App from "@/app/App.tsx";
+import reportWebVitals from "./reportWebVitals";
 
-// Import the generated route tree
-import { routeTree } from "./routeTree.gen";
+const rootElement = document.getElementById("root")!;
+const root = ReactDOM.createRoot(rootElement);
 
-import "./styles.css";
-import reportWebVitals from "./reportWebVitals.ts";
-
-// Create a new router instance
-const router = createRouter({
-	routeTree,
-	context: {},
-	defaultPreload: "intent",
-	scrollRestoration: true,
-	defaultStructuralSharing: true,
-	defaultPreloadStaleTime: 0,
-});
-
-// Register the router instance for type safety
-declare module "@tanstack/react-router" {
-	interface Register {
-		router: typeof router;
-	}
+if (import.meta.env.MODE === "development") {
+  const { worker } = await import("./__mocks__/browser");
+  worker.start();
 }
 
-// Render the app
-const rootElement = document.getElementById("app");
-if (rootElement && !rootElement.innerHTML) {
-	const root = ReactDOM.createRoot(rootElement);
-	root.render(
-		<StrictMode>
-			<RouterProvider router={router} />
-		</StrictMode>,
-	);
-}
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>
+);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
